others exceptions update and delete

+++ src/main/java/tfr/dev/tfrDSCommerce/services/ProductService.java
@@ -3,15 +3,18 @@ package tfr.dev.tfrDSCommerce.services;
 
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import tfr.dev.tfrDSCommerce.dto.ProductDTO;
 import tfr.dev.tfrDSCommerce.entities.Product;
 import tfr.dev.tfrDSCommerce.repositories.ProductRepository;
 import org.springframework.data.domain.Sort.Direction;
+import tfr.dev.tfrDSCommerce.services.exceptions.DataBaseException;
 import tfr.dev.tfrDSCommerce.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -51,23 +54,32 @@ public class ProductService {
         }
 
 
-        @Transactional
-        public ProductDTO update(Long id, ProductDTO dto){
+    public ProductDTO update(Long id, ProductDTO dto) {
         try {
-            Product entity = repo.getReferenceById(id); // não vai ao BD. É monitorado pela JPA.
+            Product entity = repo.getReferenceById(id);
             copyToDto(dto, entity);
             entity = repo.save(entity);
             return new ProductDTO(entity);
-        }catch (EntityNotFoundException e){
-            throw new ResourceNotFoundException("invalid id");
         }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+    }
 
-        }
 
-        @Transactional
-        public void delete(Long id){
-        repo.deleteById(id);
+
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repo.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
         }
+        try {
+            repo.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DataBaseException("Falha de integridade referencial");
+        }
+    }
 
       private void copyToDto(ProductDTO dto, Product entity){
         entity.setName(dto.getName());
